requirements.txt
git+https://github.com/argmaxinc/whisperkittools.git
ffmpeg-python
nltk
python-dotenv
openai
tqdm
